[DEVELOPMENT]-pure virtual func event producer
